cast to get rid of compiler warning
git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@1540 1c601a1c-44f4-4d93-90ac-4b632a93592e

+++ tee-sdk/trunk/tz/dev/tv/src/svc_vmcalls.c
@@ -68,7 +68,7 @@ int svc_utpm_unseal(void *in,
   return vmcall(TV_HC_UTPM_UNSEAL, /* eax */
                 (uint32_t)inbuf2,  /* ecx */
                 (uint32_t)outbuf2, /* edx */
-                digestAtCreation,  /* esi */
+                (uint32_t)digestAtCreation,  /* esi */
                 0);                /* edi */
 }
